Fixed panning when scale is less than minzoom.
As mentioned in issue #21, the image pans and snpas back to its own center when zoomed out (> minzoom). This is fixed now. Now, the cropper decides where to snap and pan. If the image is zoomed out such that y translation is > 0, we snap top of the image to the view. If the image is zoomed out such that bottom of the image is inside the view, it snaps to bottom of the view. In other cases, we try to maintain the zoom focus point where it is. The same goes for x dimension.

Also fixed image panning issue when scale is min zoom. Issue #10.

Apart from these issues, user can set min zoom and it won't be overridden by onLayout method. But this has its own pitfall. If the user changes the image, it is required for the user to change min zoom manually.

Signed-off-by: Jay Rambhia <jayrambhia777@gmail.com>